+++ dao/src/main/java/be/groups/glanguage/glanguage/api/dao/formula/description/combination/FormulaParameterCombinationItemDao.java
@@ -10,4 +10,4 @@ import org.springframework.stereotype.*;
 @Repository
 public interface FormulaParameterCombinationItemDao extends
                                                     JpaRepository<FormulaParameterCombinationItem,
-                                                        Integer> {}
+                                                        Long> {}
